Add REST API paths documentation to CityBuild API
A new documentation file has been created to provide further information about the accessible paths via the REST API of the CityBuild API. This includes details such as the type of request, required permissions, and the response. Additionally, "Rest API" category has been added to the sidebar for better navigation in the documentation.